submit F.cpp to 'F - Strings of Eternity' (abc135) [C++14 (GCC 5.4.1)]

+++ 2019/0727_ABC135/F.cpp
@@ -232,7 +232,7 @@ int main()
   int N = S.size();
   stringstream U{};
   U << S << S << S;
-  for (auto i = 0; i < 2 * T.size() / N; i++)
+  for (auto i = 0; i < 2 * static_cast<int>(T.size()) / N; i++)
   {
     U << S;
   }
